delete had errors, but i fixed it

+++ react-app/src/store/restaurants.js
@@ -89,8 +89,10 @@ export const destroyRestaurant = restaurantId => async dispatch => {
     //     dispatch(removeRestaurant(restaurantId));
     //     return result;
     // }
-    dispatch(removeRestaurant(restaurantId))
-    return deleted;
+    if (deleted) {
+        dispatch(removeRestaurant(restaurantId))
+        return deleted;
+    }
 };
